Split AppendingTransformerTest
