+++ src/intiTest/kotlin/com/github/jengelman/gradle/plugins/shadow/transformers/AppendingTransformerTest.kt
@@ -1,0 +1,54 @@
+package com.github.jengelman.gradle.plugins.shadow.transformers
+
+import assertk.assertThat
+import assertk.assertions.isEqualTo
+import kotlin.io.path.appendText
+import org.junit.jupiter.api.Test
+
+class AppendingTransformerTest : BaseTransformerTest() {
+  @Test
+  fun appendingTransformer() {
+    val one = buildJarOne {
+      insert(ENTRY_TEST_PROPERTIES, CONTENT_ONE)
+    }
+    val two = buildJarTwo {
+      insert(ENTRY_TEST_PROPERTIES, CONTENT_TWO)
+    }
+    projectScriptPath.appendText(
+      transform<AppendingTransformer>(
+        shadowJarBlock = fromJar(one, two),
+        transformerBlock = """
+          resource = '$ENTRY_TEST_PROPERTIES'
+        """.trimIndent(),
+      ),
+    )
+
+    run(shadowJarTask)
+
+    assertThat(outputShadowJar.getContent(ENTRY_TEST_PROPERTIES).trimIndent())
+      .isEqualTo(CONTENT_ONE_TWO)
+  }
+
+  @Test
+  fun appendingTransformerShortSyntax() {
+    val one = buildJarOne {
+      insert(ENTRY_TEST_PROPERTIES, CONTENT_ONE)
+    }
+    val two = buildJarTwo {
+      insert(ENTRY_TEST_PROPERTIES, CONTENT_TWO)
+    }
+    projectScriptPath.appendText(
+      """
+        $shadowJar {
+          ${fromJar(one, two)}
+          append('$ENTRY_TEST_PROPERTIES')
+        }
+      """.trimIndent(),
+    )
+
+    run(shadowJarTask)
+
+    assertThat(outputShadowJar.getContent(ENTRY_TEST_PROPERTIES).trimIndent())
+      .isEqualTo(CONTENT_ONE_TWO)
+  }
+}
